refactor(AppBar): extract Logo component
